% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{all_correlations_mX}
\alias{all_correlations_mX}
\title{Calculate the correlation of all the sub-models of mX and vy}
\usage{
all_correlations_mX(vy, mX, intercept_col = 1L, bIntercept = FALSE,
  bCentre = FALSE, cores = 1L)
}
\arguments{
\item{vy}{Vector of responses}

\item{mX}{Covariate matrix}

\item{intercept_col}{The index of the column in mX containing the intercept, if any}

\item{bIntercept}{Logical value indicating whether there is an intercept column or not}

\item{bCentre}{Logical value indicating whether to centre the response vector and covariance matrix or not}

\item{cores}{The number of cores to use}
}
\value{
The vector of correlations
}
\description{
Calculate the correlation of all the sub-models of mX and vy
}

